renamed http2-proxy to http2-cache
