+++ integration-test/k3po/src/test/scripts/should.support.long.pushing.rpt
@@ -9,12 +9,14 @@ read header "Host" "localhost:8080"
 read header "Upgrade" /(?i:websocket)/
 read header "Connection" /(?i:Upgrade)/
 read header "Sec-WebSocket-Key" /(?<key>[a-zA-Z0-9+\/=]{24})/
+read header "Sec-WebSocket-Protocol" 'h2'
 read header "Sec-WebSocket-Version" "13"
 
 write status "101" "Switching Protocols"
 write version "HTTP/1.1"
 write header "Upgrade" "websocket"
 write header "Connection" "Upgrade"
+write header "Sec-WebSocket-Protocol" 'h2'
 write header "Sec-WebSocket-Accept" ${ws:handshakeHash(key)}
 
 # HTTP 2 Client Connection Preface
@@ -27,17 +29,17 @@ read option mask [0x00 0x00 0x00 0x00]
 read [0x82 0x89] ([0..4] :readMask)
 read option mask ${readMask}
 read [0x00 0x00 0x00]
-read [0x04]
-read [0x00]
-read [0x00 0x00 0x00 0x00]
+     [0x04]
+     [0x00]
+     [0x00 0x00 0x00 0x00]
 read option mask [0x00 0x00 0x00 0x00]
 
 # Ack Settings
 write [0x82 0x09]
-write [0x00 0x00 0x00]
-write [0x04]
-write [0x01]
-write [0x00 0x00 0x00 0x00]
+      [0x00 0x00 0x00]
+      [0x04]
+      [0x01]
+      [0x00 0x00 0x00 0x00]
 
 # HTTP 2 Headers
 read [0x82 0x89] ([0..4] :readMask)
@@ -48,19 +50,18 @@ read [0x04]                                     # flags
 read [0x00 0x00 0x00 0x01]                      # stream id = 1
 read option mask [0x00 0x00 0x00 0x00]
 
-  # Authority
 read [0x82 0x8b] ([0..4] :readMask)
 read option mask ${readMask}
-read [0x86 0x82 0x41 0x86 0xa0 0xe4 0x1d 0x13 0x9d 0x09 0x84]
+read [0x87 0x82 0x41 0x86 0xa0 0xe4 0x1d 0x13 0x9d 0x09 0x84]
 read option mask [0x00 0x00 0x00 0x00]
 
 # Content (0 length)
 read [0x82 0x89] ([0..4] :readMask)
 read option mask ${readMask}
 read [0x00 0x00 0x00]                           # length = 0
-read [0x00]                                     # type = data
-read [0x01]
-read [0x00 0x00 0x00 0x01]                      # stream id = 1
+     [0x00]                                     # type = data
+     [0x01]
+     [0x00 0x00 0x00 0x01]                      # stream id = 1
 read option mask [0x00 0x00 0x00 0x00]
