+++ lib/configuration.js
@@ -45,7 +45,7 @@ confProto.getTransport = function (url) {
             console.log("Opening WS transport: " + url);
         }
         // TODO, maybe enable perMessageDeflate in production or on debug??
-        this._activeWSConnections[url] = websocket(url, "http2", {perMessageDeflate: false});
+        this._activeWSConnections[url] = websocket(url, "h2", {perMessageDeflate: false});
     }
     return this._activeWSConnections[url];
 };
