+++ package.json
@@ -4,7 +4,7 @@
   "description": "Exposes http caching to the browser by adding functionality to XMLHttpRequest, and then running XMLHttpRequest over http2 over WebSockets",
   "main": "index.js",
   "scripts": {
-    "preinstall" : "browserify lib/http2-proxy.js -o builds/http2-proxy.js",
+    "preinstall" : "browserify lib/http2-cache.js -o builds/http2-cache.js",
     "lint": "jshint lib/*.js test/*.js integration-test/*.js",
     "pretest": "npm run lint",
     "test": "mocha",
