+++ test/http2-xhr-test.js
@@ -1,7 +1,7 @@
 /* jshint ignore:start */
 XMLHttpRequest = require("xhr2").XMLHttpRequest;
 /* jshint ignore:end */
-require("../lib/http2-proxy");
+require("../lib/http2-cache");
 
 var assert = require('assert'),
     http = require('http'),
